API_day1/API_1day/backend/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 25.5.1
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohttp=3.7.4.post0=py39h3811e60_1
annotated-types=0.7.0=pyhd8ed1ab_1
anyio=4.9.0=pyh29332c3_0
arrow-cpp=19.0.0=h865e1df_2
async-timeout=3.0.1=py_1000
attrs=25.3.0=pyh71513ae_0
aws-c-auth=0.6.19=h5eee18b_0
aws-c-cal=0.5.20=hdbd6064_0
aws-c-common=0.8.5=h5eee18b_0
aws-c-compression=0.2.16=h5eee18b_0
aws-c-event-stream=0.2.15=h6a678d5_0
aws-c-http=0.6.25=h5eee18b_0
aws-c-io=0.13.10=h5eee18b_0
aws-c-mqtt=0.7.13=h5eee18b_0
aws-c-s3=0.1.51=hdbd6064_0
aws-c-sdkutils=0.1.6=h5eee18b_0
aws-checksums=0.1.13=h5eee18b_0
aws-crt-cpp=0.18.16=h6a678d5_0
aws-sdk-cpp=1.11.212=hecad206_0
blas=1.0=mkl
boost-cpp=1.78.0=he72f1d9_0
bottleneck=1.4.2=py39ha9d4c09_0
brotli-python=1.0.9=py39h6a678d5_9
bzip2=1.0.8=h5eee18b_6
c-ares=1.19.1=h5eee18b_0
ca-certificates=2025.6.15=hbd8a1cb_0
certifi=2025.6.15=pyhd8ed1ab_0
chardet=4.0.0=py39hf3d152e_3
charset-normalizer=3.3.2=pyhd3eb1b0_0
click=8.1.8=pyh707e725_0
colorama=0.4.6=pyhd8ed1ab_1
cpuonly=2.0=0
dataclasses=0.8=pyhc8e2a94_3
datasets=3.6.0=pyhd8ed1ab_0
dill=0.3.8=pyhd8ed1ab_0
dnspython=2.7.0=pyhff2d567_1
email-validator=2.2.0=pyhd8ed1ab_1
email_validator=2.2.0=hd8ed1ab_1
exceptiongroup=1.3.0=pyhd8ed1ab_0
expat=2.7.1=h6a678d5_0
fastapi=0.115.13=pyhe01879c_0
fastapi-cli=0.0.7=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
filelock=3.18.0=pyhd8ed1ab_0
freetype=2.13.3=h4a9f257_0
fsspec=2025.3.0=pyhd8ed1ab_0
gflags=2.2.2=he1b5a44_1004
glog=0.5.0=h48cff8f_0
gmp=6.3.0=h6a678d5_0
gnutls=3.6.15=he1e5248_0
h11=0.16.0=pyhd8ed1ab_0
h2=4.2.0=pyhd8ed1ab_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pyh29332c3_0
httptools=0.6.4=py39h8cd3c5a_0
httpx=0.28.1=pyhd8ed1ab_0
huggingface_hub=0.31.4=pyhd8ed1ab_0
hyperframe=6.1.0=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.7=py39h06a4308_0
importlib-metadata=8.7.0=pyhe01879c_1
importlib_metadata=8.7.0=h40b2b14_1
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.6=pyhd8ed1ab_0
joblib=1.5.1=pyhd8ed1ab_0
jpeg=9e=h5eee18b_3
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lcms2=2.16=h92b89f2_1
ld_impl_linux-64=2.40=h12ee557_0
lerc=4.0.0=h6a678d5_0
libabseil=20250127.0=cxx17_h6a678d5_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h166bdaf_7
libbrotlidec=1.0.9=h166bdaf_7
libbrotlienc=1.0.9=h166bdaf_7
libcblas=3.9.0=12_h832b8c9_netlib
libcurl=8.12.1=hc9e6f67_0
libdeflate=1.22=h5eee18b_0
libedit=3.1.20230828=h5eee18b_0
libev=4.33=h516909a_1
libevent=2.1.12=hdbd6064_1
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_2
libgcc-ng=15.1.0=h69a702a_2
libgfortran=15.1.0=h69a702a_2
libgfortran-ng=15.1.0=h69a702a_2
libgfortran5=15.1.0=hcea5267_2
libgomp=15.1.0=h767d61c_2
libgrpc=1.71.0=h2d74bed_0
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=12_hd37a5e2_netlib
libnghttp2=1.57.0=h2d74bed_0
libpng=1.6.39=h5eee18b_0
libprotobuf=5.29.3=h3cdef7c_1
libre2-11=2024.07.02=h6a678d5_0
libssh2=1.11.1=h251f7ec_0
libstdcxx=15.1.0=h8f9b012_2
libstdcxx-ng=15.1.0=h4852527_2
libtasn1=4.19.0=h5eee18b_0
libthrift=0.15.0=h1795dd8_2
libtiff=4.7.0=hde9077f_0
libunistring=0.9.10=h27cfd23_0
libuv=1.51.0=hb9d3cd8_0
libwebp-base=1.3.2=h5eee18b_1
libxcb=1.17.0=h9b100fa_0
lz4-c=1.9.4=h6a678d5_1
markdown-it-py=3.0.0=pyhd8ed1ab_1
markupsafe=3.0.2=py39h9399b63_1
mdurl=0.1.2=pyhd8ed1ab_1
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_2
mkl_fft=1.3.11=py39h5eee18b_0
mkl_random=1.2.8=py39h1128e8f_0
multidict=6.0.2=py39hb9d737c_1
multiprocess=0.70.12.2=py39hb9d737c_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
numexpr=2.10.1=py39h3c60e43_0
numpy=1.23.5=py39h3d75532_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.5.2=h0d4d230_1
openssl=3.5.0=h7b32b05_1
orc=2.1.1=hd396ef6_0
packaging=25.0=pyh29332c3_1
pandas=2.2.3=py39h6a678d5_0
pillow=11.1.0=py39hac6e08b_1
pip=25.1=pyhc872135_2
pthread-stubs=0.3=h0ce48e5_1
pyarrow=19.0.0=py39h6a678d5_1
pydantic=2.11.7=pyh3cfb1c2_0
pydantic-core=2.33.2=py39h20260ba_0
pygments=2.19.2=pyhd8ed1ab_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.23=he99959a_0
python-dateutil=2.9.0.post0=pyhff2d567_1
python-dotenv=1.1.0=pyh29332c3_1
python-multipart=0.0.20=pyhff2d567_0
python-tzdata=2025.2=pyhd8ed1ab_0
python-xxhash=3.0.0=py39hb9d737c_1
python_abi=3.9=2_cp39
pytorch=1.13.1=py3.9_cpu_0
pytorch-mutex=1.0=cpu
pytz=2025.2=pyhd8ed1ab_0
pyyaml=6.0=py39hb9d737c_4
re2=2024.07.02=hdb19cb5_0
readline=8.2=h5eee18b_0
regex=2022.4.24=py39hb9d737c_0
requests=2.32.4=py39h06a4308_0
rich=14.0.0=pyh29332c3_0
rich-toolkit=0.11.3=pyh29332c3_0
s2n=1.3.27=hdbd6064_0
sacremoses=0.0.53=pyhd8ed1ab_0
setuptools=78.1.1=py39h06a4308_0
shellingham=1.5.4=pyhd8ed1ab_1
six=1.17.0=pyhd8ed1ab_0
snappy=1.2.1=h6a678d5_0
sniffio=1.3.1=pyhd8ed1ab_1
sqlite=3.45.3=h5eee18b_0
starlette=0.46.2=pyh81abbef_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.14=h993c535_1
tokenizers=0.12.1=py39h312a7cb_0
torchvision=0.14.1=py39_cpu
tqdm=4.67.1=pyhd8ed1ab_1
transformers=4.26.1=pyhd8ed1ab_0
typer=0.16.0=pyh167b9f4_0
typer-slim=0.16.0=pyhe01879c_0
typer-slim-standard=0.16.0=hf964461_0
typing-extensions=4.12.2=py39h06a4308_0
typing-inspection=0.4.1=pyhd8ed1ab_0
typing_extensions=4.12.2=py39h06a4308_0
tzdata=2025b=h04d1e81_0
urllib3=2.3.0=py39h06a4308_0
utf8proc=2.6.1=h5eee18b_1
uvicorn=0.34.3=pyh31011fe_0
uvicorn-standard=0.34.3=h31011fe_0
uvloop=0.21.0=py39h8cd3c5a_1
watchfiles=1.1.0=py39he612d8f_0
websockets=15.0.1=py39h8cd3c5a_0
wheel=0.45.1=py39h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xxhash=0.8.0=h7f98852_3
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.7.2=py39hb9d737c_2
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.6=hc292b87_0
